Add files via upload
This Python script is designed to scrape data from a specific product page on Flipkart, focusing on parameters such as title, price, and rating. It utilizes the BeautifulSoup library for web scraping and pandas for handling the data. The script continuously checks the price of the product, and if it falls below a certain threshold (in this case, 2000), it sends an email notification to the specified recipient using the smtplib library. The workflow begins by importing necessary libraries and defining functions for web scraping, sending emails, and checking prices. Then, it enters a loop to continuously monitor the price, updating a CSV file with the latest data. If the price condition is met, an email notification is triggered. Finally, the script sleeps for a specified interval before checking the price again, ensuring continuous monitoring. This script can be helpful for users looking to track price changes of specific products on e-commerce platforms like Flipkart.